Move Hibernate Spatial tests to the separate module
