+++ hibernate/hibernate-orm/src/main/java/io/quarkus/qe/hibernate/data/TestDataEntity.java
@@ -8,17 +8,8 @@ import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 
 import org.hibernate.annotations.JdbcTypeCode;
-import org.hibernate.annotations.processing.Exclude;
 import org.hibernate.type.SqlTypes;
 
-// Exclude this entity from Hibernate compile-time processing.
-// When hibernate-spatial is present on the classpath, the Hibernate annotation
-// processor may try to load spatial CriteriaBuilder extensions during dev-mode
-// recompilation triggered by modifyFile(), which leads to
-// a ServiceLoader error "GeolatteSpatialCriteriaExtension not a subtype".
-// Excluding the entity avoids this processor interaction while keeping
-// the runtime behavior tested here unchanged.
-@Exclude
 @NamedQuery(name = "TestDataEntity.FindByCharacter", query = "from TestDataEntity where character = :character order by id")
 @Table(name = "test_data")
 @Entity
