+++ hibernate/hibernate-orm/src/main/resources/application.properties
@@ -17,8 +17,3 @@ quarkus.hibernate-orm.test-hql-pu.datasource=<default>
 quarkus.hibernate-orm.test-hql-pu.packages=io.quarkus.qe.hibernate.hql
 quarkus.hibernate-orm.test-hql-pu.schema-management.strategy=drop-and-create
 quarkus.hibernate-orm.test-hql-pu.sql-load-script=test-hql-pu-import-script.sql
-
-quarkus.hibernate-orm.spatial.datasource=<default>
-quarkus.hibernate-orm.spatial.packages=io.quarkus.qe.hibernate.spatial
-quarkus.hibernate-orm.spatial.schema-management.strategy=drop-and-create
-quarkus.hibernate-orm.spatial.sql-load-script=spatial-pu-import-script.sql
