+++ hibernate/hibernate-orm/src/test/resources/mysql.properties
@@ -2,4 +2,3 @@ quarkus.datasource.mysql.db-kind=mysql
 quarkus.hibernate-orm.named.datasource=mysql
 quarkus.hibernate-orm.named-2.datasource=mysql
 quarkus.hibernate-orm.test-hql-pu.datasource=mysql
-quarkus.hibernate-orm.spatial.datasource=mysql
