+++ hibernate/hibernate-spatial/src/main/java/io/quarkus/qe/hibernate/spatial/entity/geolatte/GeolattePlace.java
