+++ hibernate/hibernate-spatial/src/main/java/io/quarkus/qe/hibernate/spatial/entity/geolatte/GeolatteRegion.java
