+++ hibernate/hibernate-spatial/src/main/java/io/quarkus/qe/hibernate/spatial/entity/jts/JtsRegion.java
