+++ hibernate/hibernate-spatial/src/main/java/io/quarkus/qe/hibernate/spatial/resource/AbstractSpatialResource.java
