+++ hibernate/hibernate-spatial/src/main/java/io/quarkus/qe/hibernate/spatial/service/JtsSpatialDatasetService.java
