+++ hibernate/hibernate-spatial/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+quarkus.hibernate-orm.spatial.datasource=<default>
+quarkus.hibernate-orm.spatial.packages=io.quarkus.qe.hibernate.spatial
+quarkus.hibernate-orm.spatial.schema-management.strategy=drop-and-create
+quarkus.hibernate-orm.spatial.sql-load-script=spatial-pu-import-script.sql
