+++ hibernate/hibernate-spatial/src/main/resources/spatial-pu-import-script.sql
