+++ hibernate/hibernate-spatial/src/test/java/io/quarkus/qe/hibernate/spatial/PostgisHibernateSpatialIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.qe.hibernate;
+package io.quarkus.qe.hibernate.spatial;
 
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
