+++ hibernate/hibernate-spatial/src/test/resources/mysql.properties
@@ -1,0 +1,2 @@
+quarkus.datasource.mysql.db-kind=mysql
+quarkus.hibernate-orm.spatial.datasource=mysql
