+++ hibernate/hibernate-spatial/src/test/resources/test.properties
@@ -1,0 +1,1 @@
+ts.database.openshift.use-internal-service-as-url=true
